the lastest

+++ ft_strmap.c
@@ -6,7 +6,7 @@
 /*   By: fgracefo <fgracefo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/17 16:40:47 by fgracefo          #+#    #+#             */
-/*   Updated: 2019/09/22 18:14:24 by fgracefo         ###   ########.fr       */
+/*   Updated: 2019/09/23 15:53:06 by fgracefo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strmap(char const *s, char (*f)(char))
 {
 	char	*str;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	if (!s || !f)
